+++ eva/src/components/dashPanelBoard.vue
@@ -57,6 +57,22 @@
             :color="theme.$accent_ui_color"
           >
             <template v-slot:activator="{ on }">
+              <div
+                class="control-button theme--dark"
+                :style="{ display: 'inline-flex' }"
+                :class="{ hide_control: !editMode }"
+                v-on="on"
+                @click="loadSvg = !loadSvg"
+                v-html="svgIcon.svg"
+              />
+            </template>
+            <span>Загрузка SVG</span>
+          </v-tooltip>
+          <v-tooltip
+            bottom
+            :color="theme.$accent_ui_color"
+          >
+            <template v-slot:activator="{ on }">
               <v-icon
                 class="control-button theme--dark"
                 :style="{ color: theme.$secondary_text }"
@@ -859,6 +875,10 @@
         :permissions-from="userPermissions"
         :id-dash-from="idDashFrom"
       />
+      <modal-loading-svg
+        v-model="loadSvg"
+        @updateModalValue="loadSvg = false"
+      />
     </div>
 
     <div
@@ -911,6 +931,7 @@ import settings from '../js/componentsSettings';
 import DashFilterPanel from './dash-filter-panel/DashFilterPanel.vue';
 import { globalTockens } from '../constants/globalTockens';
 import Notifications from '@/components/notifications';
+import { svgIcon } from '@/js/svg';
 
 export default {
   name: 'DashPanelBoard',
@@ -1059,6 +1080,8 @@ export default {
       screenHeight: this.getScreenHeight(),
       allGroups: [],
       tokens: [],
+      loadSvg: false,
+      svgIcon,
     };
   },
   computed: {